inseted values inside
- sql done
- cleanup done
