--- worknest/src/app/backend/event/requirements.txt
@@ -1,5 +1,0 @@
-Flask==3.0.1
-Flask-SQLAlchemy==3.1.1
-psycopg2-binary==2.9.6
-Flask-Cors==4.0.0
-python-dotenv==1.0.0
